fixed old NamedRequests tests to use new changes
